Revert "chore(iota-sdk): split apis.rs (#2186)"
This reverts commit b180681125cd28720ab1929466e22174f859a590.
