+++ crates/iota-sdk/src/apis.rs
@@ -2,17 +2,19 @@
 // Modifications Copyright (c) 2024 IOTA Stiftung
 // SPDX-License-Identifier: Apache-2.0
 
-use std::{collections::BTreeMap, sync::Arc};
+use std::{collections::BTreeMap, future, sync::Arc, time::Instant};
 
 use fastcrypto::encoding::Base64;
 use futures::{stream, StreamExt};
 use futures_core::Stream;
 use iota_json_rpc_api::{
-    GovernanceReadApiClient, IndexerApiClient, MoveUtilsClient, ReadApiClient, WriteApiClient,
+    CoinReadApiClient, GovernanceReadApiClient, IndexerApiClient, MoveUtilsClient, ReadApiClient,
+    WriteApiClient,
 };
 use iota_json_rpc_types::{
-    Checkpoint, CheckpointId, CheckpointPage, DevInspectArgs, DevInspectResults,
-    DryRunTransactionBlockResponse, DynamicFieldPage, IotaCommittee, IotaData,
+    Balance, Checkpoint, CheckpointId, CheckpointPage, Coin, CoinPage, DelegatedStake,
+    DevInspectArgs, DevInspectResults, DryRunTransactionBlockResponse, DynamicFieldPage,
+    EventFilter, EventPage, IotaCoinMetadata, IotaCommittee, IotaData, IotaEvent,
     IotaGetPastObjectRequest, IotaLoadedChildObjectsResponse, IotaMoveNormalizedModule,
     IotaObjectDataOptions, IotaObjectResponse, IotaObjectResponseQuery, IotaPastObjectResponse,
     IotaTransactionBlockEffects, IotaTransactionBlockResponse, IotaTransactionBlockResponseOptions,
@@ -20,11 +22,15 @@ use iota_json_rpc_types::{
     TransactionFilter,
 };
 use iota_types::{
+    balance::Supply,
     base_types::{IotaAddress, ObjectID, SequenceNumber, TransactionDigest},
     dynamic_field::DynamicFieldName,
+    event::EventID,
     iota_serde::BigInt,
+    iota_system_state::iota_system_state_summary::IotaSystemStateSummary,
     messages_checkpoint::CheckpointSequenceNumber,
-    transaction::{TransactionData, TransactionKind},
+    quorum_driver_types::ExecuteTransactionRequestType,
+    transaction::{Transaction, TransactionData, TransactionKind},
 };
 use jsonrpsee::core::client::Subscription;
 
@@ -32,6 +38,8 @@ use crate::{
     error::{Error, IotaRpcResult},
     RpcClient,
 };
+
+const WAIT_FOR_LOCAL_EXECUTION_RETRY_COUNT: u8 = 3;
 
 /// The main read API structure with functions for retrieving data about
 /// different objects and transactions
@@ -44,7 +52,6 @@ impl ReadApi {
     pub(crate) fn new(api: Arc<RpcClient>) -> Self {
         Self { api }
     }
-
     /// Return a paginated response with the objects owned by the given address,
     /// or an error upon failure.
     ///
@@ -742,5 +749,542 @@ impl ReadApi {
         version: Option<BigInt<u64>>,
     ) -> IotaRpcResult<ProtocolConfigResponse> {
         Ok(self.api.http.get_protocol_config(version).await?)
+    }
+}
+
+/// Coin Read API provides the functionality needed to get information from the
+/// Iota network regarding the coins owned by an address.
+#[derive(Debug, Clone)]
+pub struct CoinReadApi {
+    api: Arc<RpcClient>,
+}
+
+impl CoinReadApi {
+    pub(crate) fn new(api: Arc<RpcClient>) -> Self {
+        Self { api }
+    }
+
+    /// Return a paginated response with the coins for the given address, or an
+    /// error upon failure.
+    ///
+    /// The coins can be filtered by `coin_type` (e.g.,
+    /// 0x168da5bf1f48dafc111b0a488fa454aca95e0b5e::usdc::USDC)
+    /// or use `None` for the default `Coin<IOTA>`.
+    ///
+    /// # Examples
+    ///
+    /// ```rust,no_run
+    /// use std::str::FromStr;
+    ///
+    /// use iota_sdk::IotaClientBuilder;
+    /// use iota_types::base_types::IotaAddress;
+    ///
+    /// #[tokio::main]
+    /// async fn main() -> Result<(), anyhow::Error> {
+    ///     let iota = IotaClientBuilder::default().build_localnet().await?;
+    ///     let address = IotaAddress::from_str("0x0000....0000")?;
+    ///     let coins = iota
+    ///         .coin_read_api()
+    ///         .get_coins(address, None, None, None)
+    ///         .await?;
+    ///     Ok(())
+    /// }
+    /// ```
+    pub async fn get_coins(
+        &self,
+        owner: IotaAddress,
+        coin_type: Option<String>,
+        cursor: Option<ObjectID>,
+        limit: Option<usize>,
+    ) -> IotaRpcResult<CoinPage> {
+        Ok(self
+            .api
+            .http
+            .get_coins(owner, coin_type, cursor, limit)
+            .await?)
+    }
+    /// Return a paginated response with all the coins for the given address, or
+    /// an error upon failure.
+    ///
+    /// This function includes all coins. If needed to filter by coin type, use
+    /// the `get_coins` method instead.
+    ///
+    /// # Examples
+    ///
+    /// ```rust,no_run
+    /// use std::str::FromStr;
+    ///
+    /// use iota_sdk::IotaClientBuilder;
+    /// use iota_types::base_types::IotaAddress;
+    ///
+    /// #[tokio::main]
+    /// async fn main() -> Result<(), anyhow::Error> {
+    ///     let iota = IotaClientBuilder::default().build_localnet().await?;
+    ///     let address = IotaAddress::from_str("0x0000....0000")?;
+    ///     let coins = iota
+    ///         .coin_read_api()
+    ///         .get_all_coins(address, None, None)
+    ///         .await?;
+    ///     Ok(())
+    /// }
+    /// ```
+    pub async fn get_all_coins(
+        &self,
+        owner: IotaAddress,
+        cursor: Option<ObjectID>,
+        limit: Option<usize>,
+    ) -> IotaRpcResult<CoinPage> {
+        Ok(self.api.http.get_all_coins(owner, cursor, limit).await?)
+    }
+
+    /// Return the coins for the given address as a stream.
+    ///
+    /// The coins can be filtered by `coin_type` (e.g.,
+    /// 0x168da5bf1f48dafc111b0a488fa454aca95e0b5e::usdc::USDC)
+    /// or use `None` for the default `Coin<IOTA>`.
+    ///
+    /// # Examples
+    ///
+    /// ```rust,no_run
+    /// use std::str::FromStr;
+    ///
+    /// use iota_sdk::IotaClientBuilder;
+    /// use iota_types::base_types::IotaAddress;
+    ///
+    /// #[tokio::main]
+    /// async fn main() -> Result<(), anyhow::Error> {
+    ///     let iota = IotaClientBuilder::default().build_localnet().await?;
+    ///     let address = IotaAddress::from_str("0x0000....0000")?;
+    ///     let coins = iota.coin_read_api().get_coins_stream(address, None);
+    ///     Ok(())
+    /// }
+    /// ```
+    pub fn get_coins_stream(
+        &self,
+        owner: IotaAddress,
+        coin_type: Option<String>,
+    ) -> impl Stream<Item = Coin> + '_ {
+        stream::unfold(
+            (
+                vec![],
+                // cursor
+                None,
+                // has_next_page
+                true,
+                coin_type,
+            ),
+            move |(mut data, cursor, has_next_page, coin_type)| async move {
+                if let Some(item) = data.pop() {
+                    Some((item, (data, cursor, /* has_next_page */ true, coin_type)))
+                } else if has_next_page {
+                    let page = self
+                        .get_coins(owner, coin_type.clone(), cursor, Some(100))
+                        .await
+                        .ok()?;
+                    let mut data = page.data;
+                    data.reverse();
+                    data.pop().map(|item| {
+                        (
+                            item,
+                            (data, page.next_cursor, page.has_next_page, coin_type),
+                        )
+                    })
+                } else {
+                    None
+                }
+            },
+        )
+    }
+
+    /// Return a list of coins for the given address, or an error upon failure.
+    ///
+    /// Note that the function selects coins to meet or exceed the requested
+    /// `amount`. If that it is not possible, it will fail with an
+    /// insufficient fund error.
+    ///
+    /// The coins can be filtered by `coin_type` (e.g.,
+    /// 0x168da5bf1f48dafc111b0a488fa454aca95e0b5e::usdc::USDC)
+    /// or use `None` to use the default `Coin<IOTA>`.
+    ///
+    /// # Examples
+    ///
+    /// ```rust,no_run
+    /// use std::str::FromStr;
+    ///
+    /// use iota_sdk::IotaClientBuilder;
+    /// use iota_types::base_types::IotaAddress;
+    ///
+    /// #[tokio::main]
+    /// async fn main() -> Result<(), anyhow::Error> {
+    ///     let iota = IotaClientBuilder::default().build_localnet().await?;
+    ///     let address = IotaAddress::from_str("0x0000....0000")?;
+    ///     let coins = iota
+    ///         .coin_read_api()
+    ///         .select_coins(address, None, 5, vec![])
+    ///         .await?;
+    ///     Ok(())
+    /// }
+    /// ```
+    pub async fn select_coins(
+        &self,
+        address: IotaAddress,
+        coin_type: Option<String>,
+        amount: u128,
+        exclude: Vec<ObjectID>,
+    ) -> IotaRpcResult<Vec<Coin>> {
+        let mut total = 0u128;
+        let coins = self
+            .get_coins_stream(address, coin_type)
+            .filter(|coin: &Coin| future::ready(!exclude.contains(&coin.coin_object_id)))
+            .take_while(|coin: &Coin| {
+                let ready = future::ready(total < amount);
+                total += coin.balance as u128;
+                ready
+            })
+            .collect::<Vec<_>>()
+            .await;
+
+        if total < amount {
+            return Err(Error::InsufficientFund { address, amount });
+        }
+        Ok(coins)
+    }
+
+    /// Return the balance for the given coin type owned by address, or an error
+    /// upon failure.
+    ///
+    /// Note that this function sums up all the balances of all the coins
+    /// matching the given coin type. By default, if `coin_type` is set to
+    /// `None`, it will use the default `Coin<IOTA>`.
+    ///
+    /// # Examples
+    ///
+    /// ```rust,no_run
+    /// use std::str::FromStr;
+    ///
+    /// use iota_sdk::IotaClientBuilder;
+    /// use iota_types::base_types::IotaAddress;
+    ///
+    /// #[tokio::main]
+    /// async fn main() -> Result<(), anyhow::Error> {
+    ///     let iota = IotaClientBuilder::default().build_localnet().await?;
+    ///     let address = IotaAddress::from_str("0x0000....0000")?;
+    ///     let balance = iota.coin_read_api().get_balance(address, None).await?;
+    ///     Ok(())
+    /// }
+    /// ```
+    pub async fn get_balance(
+        &self,
+        owner: IotaAddress,
+        coin_type: Option<String>,
+    ) -> IotaRpcResult<Balance> {
+        Ok(self.api.http.get_balance(owner, coin_type).await?)
+    }
+
+    /// Return a list of balances for each coin type owned by the given address,
+    /// or an error upon failure.
+    ///
+    /// Note that this function groups the coins by coin type, and sums up all
+    /// their balances.
+    ///
+    /// # Examples
+    ///
+    /// ```rust,no_run
+    /// use std::str::FromStr;
+    ///
+    /// use iota_sdk::IotaClientBuilder;
+    /// use iota_types::base_types::IotaAddress;
+    ///
+    /// #[tokio::main]
+    /// async fn main() -> Result<(), anyhow::Error> {
+    ///     let iota = IotaClientBuilder::default().build_localnet().await?;
+    ///     let address = IotaAddress::from_str("0x0000....0000")?;
+    ///     let all_balances = iota.coin_read_api().get_all_balances(address).await?;
+    ///     Ok(())
+    /// }
+    /// ```
+    pub async fn get_all_balances(&self, owner: IotaAddress) -> IotaRpcResult<Vec<Balance>> {
+        Ok(self.api.http.get_all_balances(owner).await?)
+    }
+
+    /// Return the coin metadata (name, symbol, description, decimals, etc.) for
+    /// a given coin type, or an error upon failure.
+    ///
+    /// # Examples
+    ///
+    /// ```rust,no_run
+    /// use iota_sdk::IotaClientBuilder;
+    /// #[tokio::main]
+    /// async fn main() -> Result<(), anyhow::Error> {
+    ///     let iota = IotaClientBuilder::default().build_localnet().await?;
+    ///     let coin_metadata = iota
+    ///         .coin_read_api()
+    ///         .get_coin_metadata("0x2::iota::IOTA".to_string())
+    ///         .await?;
+    ///     Ok(())
+    /// }
+    /// ```
+    pub async fn get_coin_metadata(
+        &self,
+        coin_type: String,
+    ) -> IotaRpcResult<Option<IotaCoinMetadata>> {
+        Ok(self.api.http.get_coin_metadata(coin_type).await?)
+    }
+
+    /// Return the total supply for a given coin type, or an error upon failure.
+    ///
+    /// # Examples
+    ///
+    /// ```rust,no_run
+    /// use iota_sdk::IotaClientBuilder;
+    ///
+    /// #[tokio::main]
+    /// async fn main() -> Result<(), anyhow::Error> {
+    ///     let iota = IotaClientBuilder::default().build_localnet().await?;
+    ///     let total_supply = iota
+    ///         .coin_read_api()
+    ///         .get_total_supply("0x2::iota::IOTA".to_string())
+    ///         .await?;
+    ///     Ok(())
+    /// }
+    /// ```
+    pub async fn get_total_supply(&self, coin_type: String) -> IotaRpcResult<Supply> {
+        Ok(self.api.http.get_total_supply(coin_type).await?)
+    }
+}
+
+/// Event API provides the functionality to fetch, query, or subscribe to events
+/// on the Iota network.
+#[derive(Clone)]
+pub struct EventApi {
+    api: Arc<RpcClient>,
+}
+
+impl EventApi {
+    pub(crate) fn new(api: Arc<RpcClient>) -> Self {
+        Self { api }
+    }
+
+    /// Return a stream of events, or an error upon failure.
+    ///
+    /// Subscription is only possible via WebSockets.
+    /// For a list of possible event filters, see [EventFilter].
+    ///
+    /// # Examples
+    ///
+    /// ```rust, no_run
+    /// use std::str::FromStr;
+    ///
+    /// use futures::StreamExt;
+    /// use iota_json_rpc_types::EventFilter;
+    /// use iota_sdk::IotaClientBuilder;
+    /// use iota_types::base_types::IotaAddress;
+    /// #[tokio::main]
+    /// async fn main() -> Result<(), anyhow::Error> {
+    ///     let iota = IotaClientBuilder::default()
+    ///         .ws_url("wss://rpc.mainnet.iota.io:443")
+    ///         .build("https://fullnode.mainnet.iota.io:443")
+    ///         .await?;
+    ///     let mut subscribe_all = iota
+    ///         .event_api()
+    ///         .subscribe_event(EventFilter::All(vec![]))
+    ///         .await?;
+    ///     loop {
+    ///         println!("{:?}", subscribe_all.next().await);
+    ///     }
+    ///     Ok(())
+    /// }
+    /// ```
+    pub async fn subscribe_event(
+        &self,
+        filter: EventFilter,
+    ) -> IotaRpcResult<impl Stream<Item = IotaRpcResult<IotaEvent>>> {
+        match &self.api.ws {
+            Some(c) => {
+                let subscription: Subscription<IotaEvent> = c.subscribe_event(filter).await?;
+                Ok(subscription.map(|item| Ok(item?)))
+            }
+            _ => Err(Error::Subscription(
+                "Subscription only supported by WebSocket client.".to_string(),
+            )),
+        }
+    }
+
+    /// Return a list of events for the given transaction digest, or an error
+    /// upon failure.
+    pub async fn get_events(&self, digest: TransactionDigest) -> IotaRpcResult<Vec<IotaEvent>> {
+        Ok(self.api.http.get_events(digest).await?)
+    }
+
+    /// Return a paginated response with events for the given event filter, or
+    /// an error upon failure.
+    ///
+    /// The ordering of the events can be set with the `descending_order`
+    /// argument. For a list of possible event filters, see [EventFilter].
+    pub async fn query_events(
+        &self,
+        query: EventFilter,
+        cursor: Option<EventID>,
+        limit: Option<usize>,
+        descending_order: bool,
+    ) -> IotaRpcResult<EventPage> {
+        Ok(self
+            .api
+            .http
+            .query_events(query, cursor, limit, Some(descending_order))
+            .await?)
+    }
+
+    /// Return a stream of events for the given event filter.
+    ///
+    /// The ordering of the events can be set with the `descending_order`
+    /// argument. For a list of possible event filters, see [EventFilter].
+    pub fn get_events_stream(
+        &self,
+        query: EventFilter,
+        cursor: Option<EventID>,
+        descending_order: bool,
+    ) -> impl Stream<Item = IotaEvent> + '_ {
+        stream::unfold(
+            (vec![], cursor, true, query),
+            move |(mut data, cursor, first, query)| async move {
+                if let Some(item) = data.pop() {
+                    Some((item, (data, cursor, false, query)))
+                } else if (cursor.is_none() && first) || cursor.is_some() {
+                    let page = self
+                        .query_events(query.clone(), cursor, Some(100), descending_order)
+                        .await
+                        .ok()?;
+                    let mut data = page.data;
+                    data.reverse();
+                    data.pop()
+                        .map(|item| (item, (data, page.next_cursor, false, query)))
+                } else {
+                    None
+                }
+            },
+        )
+    }
+}
+
+/// Quorum API that provides functionality to execute a transaction block and
+/// submit it to the fullnode(s).
+#[derive(Clone)]
+pub struct QuorumDriverApi {
+    api: Arc<RpcClient>,
+}
+
+impl QuorumDriverApi {
+    pub(crate) fn new(api: Arc<RpcClient>) -> Self {
+        Self { api }
+    }
+
+    /// Execute a transaction with a FullNode client. `request_type`
+    /// defaults to `ExecuteTransactionRequestType::WaitForLocalExecution`.
+    /// When `ExecuteTransactionRequestType::WaitForLocalExecution` is used,
+    /// but returned `confirmed_local_execution` is false, the client will
+    /// keep retry for WAIT_FOR_LOCAL_EXECUTION_RETRY_COUNT times. If it
+    /// still fails, it will return an error.
+    pub async fn execute_transaction_block(
+        &self,
+        tx: Transaction,
+        options: IotaTransactionBlockResponseOptions,
+        request_type: Option<ExecuteTransactionRequestType>,
+    ) -> IotaRpcResult<IotaTransactionBlockResponse> {
+        let (tx_bytes, signatures) = tx.to_tx_bytes_and_signatures();
+        let request_type = request_type.unwrap_or_else(|| options.default_execution_request_type());
+        let mut retry_count = 0;
+        let start = Instant::now();
+        while retry_count < WAIT_FOR_LOCAL_EXECUTION_RETRY_COUNT {
+            let response: IotaTransactionBlockResponse = self
+                .api
+                .http
+                .execute_transaction_block(
+                    tx_bytes.clone(),
+                    signatures.clone(),
+                    Some(options.clone()),
+                    Some(request_type.clone()),
+                )
+                .await?;
+
+            match request_type {
+                ExecuteTransactionRequestType::WaitForEffectsCert => {
+                    return Ok(response);
+                }
+                ExecuteTransactionRequestType::WaitForLocalExecution => {
+                    if let Some(true) = response.confirmed_local_execution {
+                        return Ok(response);
+                    } else {
+                        // If fullnode executed the cert in the network but did not confirm local
+                        // execution, it must have timed out and hence we could retry.
+                        retry_count += 1;
+                    }
+                }
+            }
+        }
+        Err(Error::FailToConfirmTransactionStatus(
+            *tx.digest(),
+            start.elapsed().as_secs(),
+        ))
+    }
+}
+
+/// Governance API provides the staking functionality.
+#[derive(Debug, Clone)]
+pub struct GovernanceApi {
+    api: Arc<RpcClient>,
+}
+
+impl GovernanceApi {
+    pub(crate) fn new(api: Arc<RpcClient>) -> Self {
+        Self { api }
+    }
+
+    /// Return a list of [DelegatedStake] objects for the given address, or an
+    /// error upon failure.
+    pub async fn get_stakes(&self, owner: IotaAddress) -> IotaRpcResult<Vec<DelegatedStake>> {
+        Ok(self.api.http.get_stakes(owner).await?)
+    }
+
+    /// Return the [IotaCommittee] information for the given `epoch`, or an
+    /// error upon failure.
+    ///
+    /// The argument `epoch` is the known epoch id or `None` for the current
+    /// epoch.
+    ///
+    /// # Examples
+    ///
+    /// ```rust,no_run
+    /// use iota_sdk::IotaClientBuilder;
+    ///
+    /// #[tokio::main]
+    /// async fn main() -> Result<(), anyhow::Error> {
+    ///     let iota = IotaClientBuilder::default().build_localnet().await?;
+    ///     let committee_info = iota.governance_api().get_committee_info(None).await?;
+    ///     Ok(())
+    /// }
+    /// ```
+    pub async fn get_committee_info(
+        &self,
+        epoch: Option<BigInt<u64>>,
+    ) -> IotaRpcResult<IotaCommittee> {
+        Ok(self.api.http.get_committee_info(epoch).await?)
+    }
+
+    /// Return the latest IOTA system state object on-chain, or an error upon
+    /// failure.
+    ///
+    /// Use this method to access system's information, such as the current
+    /// epoch, the protocol version, the reference gas price, the total
+    /// stake, active validators, and much more. See the
+    /// [IotaSystemStateSummary] for all the available fields.
+    pub async fn get_latest_iota_system_state(&self) -> IotaRpcResult<IotaSystemStateSummary> {
+        Ok(self.api.http.get_latest_iota_system_state().await?)
+    }
+
+    /// Return the reference gas price for the network, or an error upon
+    /// failure.
+    pub async fn get_reference_gas_price(&self) -> IotaRpcResult<u64> {
+        Ok(*self.api.http.get_reference_gas_price().await?)
     }
 }
