--- crates/iota-sdk/src/apis/mod.rs
@@ -1,14 +1,0 @@
-// Copyright (c) Mysten Labs, Inc.
-// Modifications Copyright (c) 2024 IOTA Stiftung
-// SPDX-License-Identifier: Apache-2.0
-
-mod coin_read;
-mod event;
-mod governance;
-mod quorum_driver;
-mod read;
-
-pub use self::{
-    coin_read::CoinReadApi, event::EventApi, governance::GovernanceApi,
-    quorum_driver::QuorumDriverApi, read::ReadApi,
-};
